+++ chrome/background.js
@@ -160,6 +160,7 @@ chrome.runtime.onMessage.addListener(function (request, sender, sendResponse) {
 chrome.runtime.onInstalled.addListener(() => {
   // Set defaults on installation
   chrome.storage.sync.set({ canaryToken: "xnlreveal" });
+  chrome.storage.sync.set({ copyToClipboard: false });
   chrome.storage.sync.set({ checkDelay: "2" });
   chrome.storage.sync.set({ waybackRegex: "" });
   chrome.storage.sync.set({ scopeType: "" });