+++ chrome/options.js
@@ -1,4 +1,5 @@
 const canaryTokenInput = document.getElementById("canaryTokenInput");
+const copyToClipboardInput = document.getElementById("copyToClipboardInput");
 const checkDelayInput = document.getElementById("checkDelayInput");
 const waybackRegexInput = document.getElementById("waybackRegexInput");
 const saveCanaryTokenButton = document.getElementById("saveButton");
@@ -27,6 +28,9 @@ document.addEventListener("DOMContentLoaded", function () {
 // Load the user's options
 chrome.storage.sync.get(["canaryToken"], (result) => {
   canaryTokenInput.value = result.canaryToken || "xnlreveal";
+});
+chrome.storage.sync.get(["copyToClipboard"], (result) => {
+  copyToClipboardInput.checked = result.copyToClipboard || false;
 });
 chrome.storage.sync.get(["checkDelay"], (result) => {
   checkDelayInput.value = result.checkDelay || "2";
@@ -112,6 +116,7 @@ saveButton.addEventListener("click", (e) => {
   e.preventDefault(); // Prevent the form from actually submitting
 
   let canaryToken = canaryTokenInput.value;
+  const copyToClipboard = copyToClipboardInput.checked;
   const checkDelay = checkDelayInput.value;
   let waybackRegex = waybackRegexInput.value;
 
@@ -129,6 +134,7 @@ saveButton.addEventListener("click", (e) => {
 
   // Save all options
   chrome.storage.sync.set({ canaryToken });
+  chrome.storage.sync.set({ copyToClipboard });
   chrome.storage.sync.set({ checkDelay });
   chrome.storage.sync.set({ waybackRegex });
   if (scopeTypeWhiteRadio.checked) {